test: extract XML output tests
